requirements.txt
bluepy==1.3.0
paho-mqtt==1.5.1
PyYAML==5.4.1
schedule==1.1.0
